*.css -> *.scss and fixed imports
