+++ src/pages/modules/Alert/Alert.scss
@@ -12,13 +12,13 @@
     -moz-transform: translateY(-40px);
     -ms-transform: translateY(-40px);
     -o-transform: translateY(-40px);
-}
-
-#alert-content {
-    padding: 20px 40px;
-    border-radius: 500px;
-    -webkit-border-radius: 500px;
-    -moz-border-radius: 500px;
-    -ms-border-radius: 500px;
-    -o-border-radius: 500px;
+    
+    #alert-content {
+        padding: 20px 40px;
+        border-radius: 500px;
+        -webkit-border-radius: 500px;
+        -moz-border-radius: 500px;
+        -ms-border-radius: 500px;
+        -o-border-radius: 500px;
+    }
 }
