+++ src/pages/modules/Avatar/Avatar.scss
