+++ src/pages/modules/Footer/submodules/Tip/Tip.scss
