+++ src/pages/modules/LoginPhase/submodules/TokenInput/TokenInput.scss
