+++ src/pages/modules/Logo/Logo.scss
