+++ src/pages/modules/Logo/Logo.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { animated, useSpring } from 'react-spring';
 
-import './Logo.css';
+import './Logo.scss';
 
 interface LProps {
   mainState?: string;
