+++ src/pages/modules/MainPhase/submodules/SearchPhase/submodules/PageProgression/PageProgression.scss
